+++ src/App.js
@@ -1,9 +1,8 @@
-import DebugView from "./components/DebugView";
+import CalculatorPage from "./components/CalculatorPage";
 import RecipesPage from "./components/RecipesPage";
 import Nav from "./components/Nav";
-import logo from "./logo.svg";
 import "./App.scss";
-import { Link, Route, useRoute } from "wouter";
+import { Route } from "wouter";
 import * as React from "react";
 
 function App() {
@@ -13,9 +12,10 @@ function App() {
 
       <Nav />
 
-      <DebugView />
       <section id="main">
-        <Route path="/calculator">Calculator</Route>
+        <Route path="/calculator">
+          <CalculatorPage />
+        </Route>
         <Route path="/recipes">
           <RecipesPage />
         </Route>