+++ src/components/RecipesTable.jsx
@@ -34,9 +34,9 @@ export default function RecipesTable({ recipes }:Props):React.MixedElement {
             {
                 recipes.map(recipe => <tr key={recipe.name}>
                     <td>{recipe.name}</td>
-                    <td>{recipe.outputPart.name}</td>
+                    <td>{recipe.outputPart.name}-{recipe.altScore}</td>
                     <td>{recipe.outputQuantity}</td>
-                    <td>{recipe.inputPart1.name}</td>
+                    <td>{recipe.inputPart1?.name}</td>
                     <td>{recipe.inputQuantity1}</td>
                     <td>{recipe.inputPart2?.name}</td>
                     <td>{recipe.inputQuantity2}</td>