+++ src/model/Recipe.js
@@ -9,8 +9,8 @@ export interface IRecipe {
   name: string;
   outputPart: Part;
   outputQuantity: number;
-  inputPart1: Part;
-  inputQuantity1: number;
+  inputPart1?: Part;
+  inputQuantity1?: number;
   inputPart2?: Part;
   inputQuantity2?: number;
   inputPart3?: Part;
@@ -19,18 +19,18 @@ export interface IRecipe {
   inputQuantity4?: number;
   byproductPart?: Part;
   byproductQuantity?: number;
-  stage?: number;
+  stage: number;
   building?: string;
   isAlternate: boolean;
-  altScore?: number;
+  altScore: number;
 }
 
 export default class Recipe {
   name: string;
   outputPart: Part;
   outputQuantity: number;
-  inputPart1: Part;
-  inputQuantity1: number;
+  inputPart1: ?Part;
+  inputQuantity1: ?number;
   inputPart2: ?Part;
   inputQuantity2: ?number;
   inputPart3: ?Part;
@@ -39,11 +39,12 @@ export default class Recipe {
   inputQuantity4: ?number;
   byproductPart: ?Part;
   byproductQuantity: ?number;
-  stage: ?number;
+  stage: number = 0;
   building: ?string;
   isAlternate: ?boolean;
-  altScore: ?number;
+  altScore: number = 0;
   static findAll: () => Array<Recipe>;
+  static findAlternateRecipes: () => Array<Recipe>;
 
   constructor(props: IRecipe) {
     if (props) {
@@ -53,38 +54,40 @@ export default class Recipe {
 }
 
 Recipe.findAll = moize(findAll);
+Recipe.findAlternateRecipes = moize(findAlternateRecipes);
 
-const recipe1 = new Recipe({
-  name: "floog",
-  outputPart: new Part({ name: "schmarf" }),
-  outputQuantity: 1,
-  inputPart1: new Part({ name: "larbo" }),
-  inputQuantity1: 4,
-});
+const FORBIDDEN_RECIPES = ["Recycled Rubber", "Recycled Plastic"];
 
 function findAll() {
   const data = rfdc(rawRecipes);
-  const recipes = rawRecipes.map(
-    (rawRecipe) =>
-      new Recipe({
-        name: rawRecipe.Recipe,
-        outputPart: new Part({ name: rawRecipe["Output"] }),
-        outputQuantity: Number(rawRecipe["Output qty/min"]),
-        inputPart1: new Part({ name: rawRecipe["Item 1"] }),
-        inputQuantity1: Number(rawRecipe["Q1"]),
-        inputPart2: new Part({ name: rawRecipe["Item 2"] }),
-        inputQuantity2: Number(rawRecipe["Q2"]),
-        inputPart3: new Part({ name: rawRecipe["Item 3"] }),
-        inputQuantity3: Number(rawRecipe["Q3"]),
-        inputPart4: new Part({ name: rawRecipe["Item 4"] }),
-        inputQuantity4: Number(rawRecipe["Q4"]),
-        byproductPart: new Part({ name: rawRecipe["Byproduct"] }),
-        byproductQuantity: Number(rawRecipe["QB"]),
-        isAlternate: rawRecipe["Alternate"] === "yes",
-        altScore: Number(rawRecipe["Alt Score"]),
-        stage: Number(rawRecipe["Stage"]),
-        building: rawRecipe["Building"],
-      })
-  );
+  const recipes = data
+    .filter((recipe) => FORBIDDEN_RECIPES.indexOf(recipe.Recipe) < 0) // TODO: figure out a way to detect circular dependencies without a forbidden recipe list
+    .map(
+      (rawRecipe) =>
+        new Recipe({
+          name: rawRecipe.Recipe,
+          outputPart: new Part({ name: rawRecipe["Output"] }),
+          outputQuantity: Number(rawRecipe["Output qty/min"]),
+          inputPart1: new Part({ name: rawRecipe["Item 1"] }),
+          inputQuantity1: Number(rawRecipe["Q1"]),
+          inputPart2: new Part({ name: rawRecipe["Item 2"] }),
+          inputQuantity2: Number(rawRecipe["Q2"]),
+          inputPart3: new Part({ name: rawRecipe["Item 3"] }),
+          inputQuantity3: Number(rawRecipe["Q3"]),
+          inputPart4: new Part({ name: rawRecipe["Item 4"] }),
+          inputQuantity4: Number(rawRecipe["Q4"]),
+          byproductPart: new Part({ name: rawRecipe["Byproduct"] }),
+          byproductQuantity: Number(rawRecipe["QB"]),
+          isAlternate: rawRecipe["Alternate"] === "yes",
+          altScore: Number(rawRecipe["Alt Score"] || 0),
+          stage: Number(rawRecipe["Stage"]),
+          building: rawRecipe["Building"],
+        })
+    );
   return recipes;
+}
+
+function findAlternateRecipes() {
+  const recipes = findAll();
+  return recipes.filter((recipe: Recipe) => recipe.isAlternate);
 }